refactor: move features/mcp-server -> entrypoints/mcp-server
