+++ src/commands/mcp/mcp.command.ts
@@ -9,8 +9,8 @@ import {
   runMcpOpenApis,
   runMcpProbe,
 } from '../../features/liferay-mcp/liferay-mcp.js';
-import {formatMcpDoctor, runMcpDoctor} from '../../features/mcp-server/mcp-server-doctor.js';
-import type {McpTool} from '../../features/mcp-server/mcp-server-setup.js';
+import {formatMcpDoctor, runMcpDoctor} from '../../entrypoints/mcp-server/mcp-server-doctor.js';
+import type {McpTool} from '../../entrypoints/mcp-server/mcp-server-setup.js';
 
 type McpAuthCommandOptions = {
   authorizationHeader?: string;
