+++ src/entrypoints/dashboard/dashboard-data.ts
@@ -2,7 +2,7 @@ import path from 'node:path';
 
 import fs from 'fs-extra';
 
-import {MCP_SETUP_TOOLS, type McpSetupTool, resolveMcpConfigPath} from '../../features/mcp-server/mcp-server-setup.js';
+import {MCP_SETUP_TOOLS, type McpSetupTool, resolveMcpConfigPath} from '../mcp-server/mcp-server-setup.js';
 import {
   collectDashboardWorktrees,
   type CollectDashboardWorktreeOptions,
