+++ src/entrypoints/mcp-server/mcp-server-doctor.ts
