+++ src/entrypoints/mcp-server/mcp-server-setup.ts
