+++ src/entrypoints/mcp-server/mcp-server.ts
