+++ src/entrypoints/mcp-server/tools/tool-ldev-context.ts
@@ -1,5 +1,5 @@
 import type {AppConfig} from '../../../core/config/schema.js';
-import {runAgentContext} from '../../agent/agent-context.js';
+import {runAgentContext} from '../../../features/agent/agent-context.js';
 import {runJsonTool} from './tool-result.js';
 
 export const TOOL_NAME = 'ldev_context';
