+++ src/entrypoints/mcp-server/tools/tool-result.ts
